feat(cart): verify customer before exec

+++ apps/api/src/cart/cart.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { ProductsService } from 'src/products/products.service';
+import { CustomersService } from 'src/customers/customers.service';
 import { RedisService } from 'src/redis/redis.service';
 import { Cart } from './entities/cart.entity';
 
@@ -9,51 +10,60 @@ export class CartService {
   constructor(
     private readonly redisService: RedisService,
     private readonly productsService: ProductsService,
+    private readonly customersService: CustomersService,
   ) {}
 
-  get(customerId: string) {
+  async get(customerId: string) {
+    const customer = await this.customersService.findOne(customerId);
+    if (!customer) throw new BadRequestException();
     return this.redisService.hGetOrSetJson<Cart>(
       this.name,
-      customerId,
+      customer.id,
       new Cart(),
     );
   }
 
   async add(customerId: string, productId: string, amount: number = 1) {
+    const customer = await this.customersService.findOne(customerId);
+    if (!customer) throw new BadRequestException();
     let cart = await this.redisService.hGetOrSetJson<Cart>(
       this.name,
-      customerId,
+      customer.id,
       new Cart(),
     );
     if (amount < 1) return cart;
     if (!cart.addItem) cart = new Cart(cart.items);
     cart.addItem(productId, amount);
-    await this.redisService.hSetJson(this.name, customerId, cart);
+    await this.redisService.hSetJson(this.name, customer.id, cart);
     return cart;
   }
 
   async remove(customerId: string, key: string, amount: number = 1) {
+    const customer = await this.customersService.findOne(customerId);
+    if (!customer) throw new BadRequestException();
     let cart = await this.redisService.hGetOrSetJson<Cart>(
       this.name,
-      customerId,
+      customer.id,
       new Cart(),
     );
     if (amount < 1) return cart;
     if (!cart.removeItem) cart = new Cart(cart.items);
     cart.removeItem(key, amount);
-    this.redisService.hSetJson(this.name, customerId, cart);
+    this.redisService.hSetJson(this.name, customer.id, cart);
     return cart;
   }
 
   async delete(customerId: string, key: string) {
+    const customer = await this.customersService.findOne(customerId);
+    if (!customer) throw new BadRequestException();
     let cart = await this.redisService.hGetOrSetJson<Cart>(
       this.name,
-      customerId,
+      customer.id,
       new Cart(),
     );
     if (!cart.deleteItem) cart = new Cart(cart.items);
     cart.deleteItem(key);
-    this.redisService.hSetJson(this.name, customerId, cart);
+    this.redisService.hSetJson(this.name, customer.id, cart);
     return cart;
   }
